requirements.txt
chromadb==1.5.5
numpy==2.4.1
openai==2.28.0
pandas==2.3.3
pdfplumber==0.11.9
plotly==6.1.2
python-dotenv==1.2.2
sentence-transformers==5.3.0
streamlit==1.55.0
tiktoken==0.12.0
